requirements/base.txt
Django
psycopg2
django-environ
djangorestframework

